// web/src/generated/api/models/messagingChannel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */

export type MessagingChannel = typeof MessagingChannel[keyof typeof MessagingChannel];


export const MessagingChannel = {
  SMS: 'SMS',
  WHATSAPP: 'WHATSAPP',
} as const;